fix(feature-backend-admin-scraping-entity) : duration 타입 정합화

+++ backend/src/main/java/kr/co/carrer/admin/scraping/entity/ScrapingPipeline.java
@@ -51,7 +51,7 @@ public class ScrapingPipeline {
     private ZonedDateTime lastFailedAt;
 
     @Column(name = "last_duration_ms")
-    private Long lastDurationMs;
+    private Integer lastDurationMs;
 
     @Column(name = "last_total_count")
     private Integer lastTotalCount;
